delete/get: offer more control for these commands.
get can now be used to fetch multiple objects into a (specified)
directory.  Failures can be ignored.  This is primarily for stress testing.

delete can now be used to delete a range of objects, or use the results
of a query.  No backend requirements for this change.

Various minor cleanups in this change.