Add support for the BPG side of stability score and improve documentation.
We now document how to integrate morgue into your CI process for stability
score, and offer a `morgue stability create-metric` command, for example:

```
./bin/morgue.js stability create-metric \
--project cts \
--metric-group stability \
--name test-metric \
--attribute version,1.0
```

Which simply gives:

```
Metric created
```

This is the BPG half of the stability flow, separated into two commands to
preserve the ability to allow API integrations to occur. This doesn't implement
anything beyond the minimum necessary for CI.  In addition to adding the above
command, the readme contains a walkthrough of how to use these commands together
on new releases.  The frontend is currently required for more complex
operations.