Base class for service clients
This provides the ability to inject default querystring parametersto set the
universe and project, among other things; a helper method to do requests
against a service routed at a specific root url with Coroner auth; and a
page token pager that matches the Rust services.

To use, inherit from it, and add your methods for your service.  We'll probably
extend it in future.