bpgObjectFind call can set a default fields value that is invalide given vals input.
Summary:
Make fix for invalid default fields value inside of bpgObjectFind function. When no fields value is provided, the default value is set to ["id"] or ["pid"], if the input vals is not an array, this is invalid. Fix to default fields to "id" or ["id"] depending on if vals is an array or not.

T22827

Test Plan: none. Manually re-ran `morgue retention list` and it is fixed with this change.

Reviewers: mrashkovsky

Subscribers: sfollari

Tags: #squad_product

Differential Revision: https://hack.in.backtrace.io/D6182