morgue: use server side delete by query
The `morgue delete` command can take a query, rather than a list of
objects, as input. This is implemented by first querying coronerd for the
list of objects and then making separate requests to delete the individual
objects in fixed sized batches.

As of Jan 2023, coronerd supports passing a query directly to the
api/delete endpoint which can be more efficient than the back and forth
described above. However, morgue was not updated to take advantage of
this.

This commit fixes this and uses the api/delete endpoint directly with
a filter-only query (no folds, etc). Updates the documentation to provide
some example usage.

Internal ref: BT-2817