morgue/sampling: adds last transaction identifier to sampling status output
In coronerd 1.62.50 and later, the status for a sampling group will
contain the last transaction identifier (txid) for an object that was
accepted by that sampling group. Having the txid can aid in tracking down
objects that were accepted based on a particular sampling group but were
then reprocessed and moved into a different group.

This commit adds the last accepted txid to the output in
`samplingStatusProject` with a fallback to "unk" if the txid is not
available.

Internal ref BT-3250