morgue/view: search for a view using the "%dashboard%" prefix
The names of views created in the Backtrace dashboard are prefixed with
"%dashboard%". To hide this internal detail and to make the command
"morgue view delete" more convenient to use, automatically search for a
view including the "%dashboard%" prefix if a view cannot be found with that
name.

Addresses a problem reported where a "morgue view delete" command would
return "Error: View not found." unexpectedly.

Internal Ref BT-5397